requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.2
dateparser==1.1.0
Flask==1.1.2
Flask-Cors==3.0.10
gate-api==4.22.2
greenlet==1.1.2
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
multidict==5.2.0
psycopg2-binary==2.9.1
pyaml==21.8.3
pycodestyle==2.7.0
PyPika==0.48.8
python-binance==1.0.15
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
redis==3.5.3
regex==2021.10.8
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.26
toml==0.10.2
typing-extensions==3.10.0.2
tzlocal==3.0
ujson==4.2.0
urllib3==1.26.7
uWSGI==2.0.20
websockets==9.1
Werkzeug==2.0.2
yarl==1.7.0
zipp==3.6.0
